Fix empty-relevant_cols dtype crash in _solve_branch_dp
relevant_cols is legitimately empty for a scene with no analytic-IK
projections at all (e.g. pyrobosim's pick_place_task, whose Pick/Place
targets are plain FK-residual equality constraints, not ProjOperators) --
jnp.asarray([]) defaults to float64 with nothing to infer a dtype from,
and x0[cols] then raises "Indexer must have integer or boolean type, got
indexer with type float64". Fixed by pinning dtype=int32 explicitly.

Found validating SmallContinuousVRPSolver's Boltzmann-sampled assignment
against pick_place_task -- the first real scene with genuine assignment
freedom (n_variables=2, n_agents=2, from PCOP leaving Pick/Place's `r:
Robot` parameter unresolved). With the fix: solves end-to-end through the
production path, and a fresh population's sampled assign combos cover all
4 possible (agent-per-variable) choices -- real diversity, not collapse to
one, on real assignment freedom rather than only the no-op (n_variables=0)
case every scene built before this had.

+++ src/goc_mpc/evolutionary_waypoint_solver/small_continuous_vrp.py
@@ -285,7 +285,16 @@ def _solve_branch_dp(problem, chain_order, chain_node_entries, relevant_cols, wp
     `_node_candidates`' `itertools.product` enumeration -- both use
     default C/row-major order, so they agree)."""
     pop = wp0.shape[0]
-    cols = jnp.asarray(relevant_cols)
+    # dtype=int32 explicitly: relevant_cols is empty for a scene with no
+    # projections at all (e.g. pick_place_task, whose Pick/Place targets
+    # are plain FK-residual equality constraints, not analytic-IK
+    # ProjOperators) -- jnp.asarray([]) defaults to float64 with nothing
+    # else to infer a dtype from, and float-indexing x0[cols] below raises.
+    # An empty int array here is correct either way: every node then has
+    # entries=[] too (no projection anywhere in the problem), so
+    # _node_candidates' own K=1 trivial-candidate path is what actually
+    # makes the whole DP a no-op, not this array's shape.
+    cols = jnp.asarray(relevant_cols, dtype=jnp.int32)
     rows_by_node = []
     for nid in chain_order:
         entries = chain_node_entries.get(nid, [])
